refactor(NODE-3356): Update command monitoring logging (#2853)
