+++ lib/core/connection/apm.js
@@ -1,128 +1,15 @@
 'use strict';
-const Msg = require('../connection/msg').Msg;
 const KillCursor = require('../connection/commands').KillCursor;
 const GetMore = require('../connection/commands').GetMore;
 const calculateDurationInMs = require('../../utils').calculateDurationInMs;
-
-/** Commands that we want to redact because of the sensitive nature of their contents */
-const SENSITIVE_COMMANDS = new Set([
-  'authenticate',
-  'saslStart',
-  'saslContinue',
-  'getnonce',
-  'createUser',
-  'updateUser',
-  'copydbgetnonce',
-  'copydbsaslstart',
-  'copydb'
-]);
-
-const HELLO_COMMANDS = new Set(['hello', 'ismaster', 'isMaster']);
+const extractCommand = require('../../command_utils').extractCommand;
 
 // helper methods
-const extractCommandName = commandDoc => Object.keys(commandDoc)[0];
 const namespace = command => command.ns;
 const databaseName = command => command.ns.split('.')[0];
-const collectionName = command => command.ns.split('.')[1];
 const generateConnectionId = pool =>
   pool.options ? `${pool.options.host}:${pool.options.port}` : pool.address;
-const maybeRedact = (commandName, cmd, result) =>
-  SENSITIVE_COMMANDS.has(commandName) ||
-  (HELLO_COMMANDS.has(commandName) && cmd.speculativeAuthenticate)
-    ? {}
-    : result;
 const isLegacyPool = pool => pool.s && pool.queue;
-
-const LEGACY_FIND_QUERY_MAP = {
-  $query: 'filter',
-  $orderby: 'sort',
-  $hint: 'hint',
-  $comment: 'comment',
-  $maxScan: 'maxScan',
-  $max: 'max',
-  $min: 'min',
-  $returnKey: 'returnKey',
-  $showDiskLoc: 'showRecordId',
-  $maxTimeMS: 'maxTimeMS',
-  $snapshot: 'snapshot'
-};
-
-const LEGACY_FIND_OPTIONS_MAP = {
-  numberToSkip: 'skip',
-  numberToReturn: 'batchSize',
-  returnFieldsSelector: 'projection'
-};
-
-const OP_QUERY_KEYS = [
-  'tailable',
-  'oplogReplay',
-  'noCursorTimeout',
-  'awaitData',
-  'partial',
-  'exhaust'
-];
-
-/**
- * Extract the actual command from the query, possibly upconverting if it's a legacy
- * format
- *
- * @param {Object} command the command
- */
-const extractCommand = command => {
-  if (command instanceof GetMore) {
-    return {
-      getMore: command.cursorId,
-      collection: collectionName(command),
-      batchSize: command.numberToReturn
-    };
-  }
-
-  if (command instanceof KillCursor) {
-    return {
-      killCursors: collectionName(command),
-      cursors: command.cursorIds
-    };
-  }
-
-  if (command instanceof Msg) {
-    return command.command;
-  }
-
-  if (command.query && command.query.$query) {
-    let result;
-    if (command.ns === 'admin.$cmd') {
-      // upconvert legacy command
-      result = Object.assign({}, command.query.$query);
-    } else {
-      // upconvert legacy find command
-      result = { find: collectionName(command) };
-      Object.keys(LEGACY_FIND_QUERY_MAP).forEach(key => {
-        if (typeof command.query[key] !== 'undefined')
-          result[LEGACY_FIND_QUERY_MAP[key]] = command.query[key];
-      });
-    }
-
-    Object.keys(LEGACY_FIND_OPTIONS_MAP).forEach(key => {
-      if (typeof command[key] !== 'undefined') result[LEGACY_FIND_OPTIONS_MAP[key]] = command[key];
-    });
-
-    OP_QUERY_KEYS.forEach(key => {
-      if (command[key]) result[key] = command[key];
-    });
-
-    if (typeof command.pre32Limit !== 'undefined') {
-      result.limit = command.pre32Limit;
-    }
-
-    if (command.query.$explain) {
-      return { explain: result };
-    }
-
-    return result;
-  }
-
-  return command.query ? command.query : command;
-};
 
 const extractReply = (command, reply) => {
   if (command instanceof GetMore) {
@@ -183,15 +70,15 @@ class CommandStartedEvent {
    * @param {Object} command the command
    */
   constructor(pool, command) {
-    const cmd = extractCommand(command);
-    const commandName = extractCommandName(cmd);
+    const extractedCommand = extractCommand(command);
+    const commandName = extractedCommand.name;
     const connectionDetails = extractConnectionDetails(pool);
 
     Object.assign(this, connectionDetails, {
       requestId: command.requestId,
       databaseName: databaseName(command),
       commandName,
-      command: maybeRedact(commandName, cmd, cmd)
+      command: extractedCommand.shouldRedact ? {} : extractedCommand.cmd
     });
   }
 }
@@ -207,15 +94,15 @@ class CommandSucceededEvent {
    * @param {Array} started a high resolution tuple timestamp of when the command was first sent, to calculate duration
    */
   constructor(pool, command, reply, started) {
-    const cmd = extractCommand(command);
-    const commandName = extractCommandName(cmd);
+    const extractedCommand = extractCommand(command);
+    const commandName = extractedCommand.name;
     const connectionDetails = extractConnectionDetails(pool);
 
     Object.assign(this, connectionDetails, {
       requestId: command.requestId,
       commandName,
       duration: calculateDurationInMs(started),
-      reply: maybeRedact(commandName, cmd, extractReply(command, reply))
+      reply: extractedCommand.shouldRedact ? {} : extractReply(command, reply)
     });
   }
 }
@@ -231,15 +118,15 @@ class CommandFailedEvent {
    * @param {Array} started a high resolution tuple timestamp of when the command was first sent, to calculate duration
    */
   constructor(pool, command, error, started) {
-    const cmd = extractCommand(command);
-    const commandName = extractCommandName(cmd);
+    const extractedCommand = extractCommand(command);
+    const commandName = extractedCommand.name;
     const connectionDetails = extractConnectionDetails(pool);
 
     Object.assign(this, connectionDetails, {
       requestId: command.requestId,
       commandName,
       duration: calculateDurationInMs(started),
-      failure: maybeRedact(commandName, cmd, error)
+      failure: extractedCommand.shouldRedact ? {} : error
     });
   }
 }
