+++ lib/core/connection/connection.js
@@ -254,10 +254,10 @@ class Connection extends EventEmitter {
     // Debug Log
     if (this.logger.isDebug()) {
       if (!Array.isArray(buffer)) {
-        this.logger.debug(`writing buffer [${buffer.toString('hex')}] to ${this.address}`);
+        this.logger.debug(`writing buffer [ ${buffer.length} ] to ${this.address}`);
       } else {
         for (let i = 0; i < buffer.length; i++)
-          this.logger.debug(`writing buffer [${buffer[i].toString('hex')}] to ${this.address}`);
+          this.logger.debug(`writing buffer [ ${buffer[i].length} ] to ${this.address}`);
       }
     }
