+++ lib/core/connection/pool.js
@@ -390,8 +390,8 @@ function messageHandler(self) {
     if (self.logger.isDebug()) {
       self.logger.debug(
         f(
-          'message [%s] received from %s:%s',
-          message.raw.toString('hex'),
+          'message [ %s ] received from %s:%s',
+          message.raw.length,
           self.options.host,
           self.options.port
         )
