+++ lib/core/topologies/server.js
@@ -16,6 +16,7 @@ var inherits = require('util').inherits,
   createCompressionInfo = require('./shared').createCompressionInfo,
   resolveClusterTime = require('./shared').resolveClusterTime,
   SessionMixins = require('./shared').SessionMixins,
+  extractCommand = require('../../command_utils').extractCommand,
   relayEvents = require('../utils').relayEvents;
 
 const collationNotSupported = require('../utils').collationNotSupported;
@@ -608,18 +609,20 @@ Server.prototype.command = function(ns, cmd, options, callback) {
   options = Object.assign({}, options, { wireProtocolCommand: false });
 
   // Debug log
-  if (self.s.logger.isDebug())
+  if (self.s.logger.isDebug()) {
+    const extractedCommand = extractCommand(cmd);
     self.s.logger.debug(
       f(
         'executing command [%s] against %s',
         JSON.stringify({
           ns: ns,
-          cmd: cmd,
+          cmd: extractedCommand.shouldRedact ? `${extractedCommand.name} details REDACTED` : cmd,
           options: debugOptions(debugFields, options)
         }),
         self.name
       )
     );
+  }
 
   // If we are not connected or have a disconnectHandler specified
   if (disconnectHandler(self, 'command', ns, cmd, options, callback)) return;
