+++ lib/operations/command.js
@@ -8,6 +8,7 @@ const handleCallback = require('../utils').handleCallback;
 const MongoError = require('../core').MongoError;
 const ReadPreference = require('../core').ReadPreference;
 const MongoDBNamespace = require('../utils').MongoDBNamespace;
+const extractCommand = require('../command_utils').extractCommand;
 
 const debugFields = [
   'authSource',
@@ -95,9 +96,10 @@ class CommandOperation extends OperationBase {
 
     // Debug information
     if (db.s.logger.isDebug()) {
+      const extractedCommand = extractCommand(command);
       db.s.logger.debug(
         `executing command ${JSON.stringify(
-          command
+          extractedCommand.shouldRedact ? `${extractedCommand.name} details REDACTED` : command
         )} against ${dbName}.$cmd with options [${JSON.stringify(
           debugOptions(debugFields, options)
         )}]`
