+++ lib/operations/command_v2.js
@@ -9,6 +9,7 @@ const maxWireVersion = require('../core/utils').maxWireVersion;
 const decorateWithExplain = require('../utils').decorateWithExplain;
 const commandSupportsReadConcern = require('../core/sessions').commandSupportsReadConcern;
 const MongoError = require('../core/error').MongoError;
+const extractCommand = require('../command_utils').extractCommand;
 
 const SUPPORTS_WRITE_CONCERN_AND_COLLATION = 5;
 
@@ -89,7 +90,12 @@ class CommandOperationV2 extends OperationBase {
     }
 
     if (this.logger && this.logger.isDebug()) {
-      this.logger.debug(`executing command ${JSON.stringify(cmd)} against ${this.ns}`);
+      const extractedCommand = extractCommand(cmd);
+      this.logger.debug(
+        `executing command ${JSON.stringify(
+          extractedCommand.shouldRedact ? `${extractedCommand.name} details REDACTED` : cmd
+        )} against ${this.ns}`
+      );
     }
 
     server.command(this.ns.toString(), cmd, this.options, (err, result) => {
